A-Packages: Reorganize into multiple packages (stretch goal)
- Organize classes into logical packages:
  - zbot.command - CommandType
  - zbot.task - Task, TaskType, Todo, Deadline, Event
  - zbot.parser - Parser
  - zbot.storage - Storage
  - zbot.ui - Ui
  - zbot.tasklist - TaskList
  - zbot - main Zbot class
- Add proper import statements for cross-package references
- Maintain src/main/java as source root
- All functionality preserved and tested
