+++ data/zbot.txt
@@ -8,3 +8,4 @@ E | 0 | birthday party | tomorrow evening | late night
 T | 1 | test OOP refactor
 D | 0 | submit project | 15/12/2019 2300
 T | 1 | test packages working
+T | 0 | test multi-package
