+++ src/main/java/zbot/Zbot.java
@@ -1,5 +1,12 @@
 package zbot;
 
+import zbot.command.CommandType;
+import zbot.parser.Parser;
+import zbot.storage.Storage;
+import zbot.task.*;
+import zbot.tasklist.TaskList;
+import zbot.ui.Ui;
+
 public class Zbot {
     private TaskList tasks;
     private Storage storage;
