+++ src/main/java/zbot/command/CommandType.java
@@ -1,4 +1,4 @@
-package zbot;
+package zbot.command;
 
 public enum CommandType {
     BYE, LIST, MARK, UNMARK, DELETE, TODO, DEADLINE, EVENT, UNKNOWN;
