+++ src/main/java/zbot/parser/Parser.java
@@ -1,4 +1,6 @@
-package zbot;
+package zbot.parser;
+
+import zbot.command.CommandType;
 
 public class Parser {
     public static CommandType parseCommand(String input) {
