+++ src/main/java/zbot/storage/Storage.java
@@ -1,5 +1,6 @@
-package zbot;
+package zbot.storage;
 
+import zbot.task.*;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.BufferedReader;
