+++ src/main/java/zbot/task/Event.java
@@ -1,4 +1,4 @@
-package zbot;
+package zbot.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
