+++ src/main/java/zbot/task/TaskType.java
@@ -1,4 +1,4 @@
-package zbot;
+package zbot.task;
 
 public enum TaskType {
     TODO, DEADLINE, EVENT;
