+++ src/main/java/zbot/task/Todo.java
@@ -1,4 +1,4 @@
-package zbot;
+package zbot.task;
 
 public class Todo extends Task {
     public Todo(String description) {
