+++ src/main/java/zbot/tasklist/TaskList.java
@@ -1,5 +1,6 @@
-package zbot;
+package zbot.tasklist;
 
+import zbot.task.Task;
 import java.util.ArrayList;
 
 public class TaskList {
